poprawki w dodawaniu ocen i czyszczenie kodu ze zbędnych klas

+++ app/src/main/java/com/projekt/dzienniczek/ui/ocenyUczen/OcenyUczenViewModel.kt
@@ -27,7 +27,7 @@ class OcenyUczenViewModel : ViewModel() {
                 if (!document.isEmpty) {
                     val list = emptyList<Subject>().toMutableList()
                     document?.forEach {
-                        list.add(it.toObject(Subject::class.java).apply { id = it.id })
+                        list.add(it.toObject(Subject::class.java).apply { id = it.id.toLong() })
                     }
                     subject.value = list
                 } else {
